backend/requirements.txt
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-extensions==3.2.3
django-filter==23.5
django-sesame==3.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
hyperlink==21.0.0
identify==2.5.35
idna==3.6
incremental==22.10.0
inflection==0.5.1
Markdown==3.5.2
nodeenv==1.8.0
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
websockets==12.0
zope.interface==6.2
